Work-around for possible crash when sharding empty files
It seems the first file of the ColReader already triggers a close of the ColReader while the second file hasn't had the time to encounter EOF yet. The Close of the ColReader closes the second file but the go-function isn't stopped.

Proper way to work around this might be an extra channel in the LineReader struct that tells the go-function in Lines() whether the backing file has been closed or not. Current implementation is more of a work-around for not crashing when this edge case occurs.